requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib